feat(lexer): improve lexer structure
